Add NPC armor rolls.
